Refactor Prediction class hierarchy
Use an Extraction superclass with Text and Page properties.

Prediction
└── Extraction
    ├── DocumentExtraction
    └── FormExtraction
